back-end/store/requirements.txt
django
djangorestframework
django-cors-headers
requests
celery
redis
psycopg2-binary
django-jazzmin